restructured the restructure
