+++ app/jobs/repository.py
@@ -2,14 +2,11 @@ from app.models.job import Job, JobStatus
 from app.extensions import db
 
 class Repository:
-    def __init__(self):
-        pass
-
     def get_jobs(self):
         return  db.session.query(Job).all()
 
     def get_job(self, job_id):
-        return db.session.query(Job).filter_by(id=job_id).one()
+        return db.session.query(Job).filter_by(id=job_id).first()
     
     def add_job(self, job: Job):
         db.session.add(job)
