+++ app/jobs/service.py
@@ -5,18 +5,24 @@ from apscheduler.triggers.cron import CronTrigger
 from flask import jsonify
 from sqlalchemy.ext.declarative import DeclarativeMeta
 
-from app.database.repository import Repository
+from app.jobs.repository import Repository
 from app.models.job import Job, JobStatus
-from app.services.scheduler_service import SchedulerService
+
+from app.extensions import scheduler
 
 repo = Repository()
-scheduler_service = SchedulerService()
+
+def execute_script(script_name):
+    script_path = os.path.join("app", script_name)
+    result = subprocess.run(
+        ["python", script_path],
+        capture_output=True,
+        text=True
+    )
+    print(result.stdout)
 
 class JobService:
-    def __init__(self):
-        pass
-
-    def get_jobs(self) -> list[Job]:
+    def get_all(self) -> list[Job]:
         jobs = repo.get_jobs()
         jobs_serialized = [
             {
@@ -29,7 +35,7 @@ class JobService:
         ] # TODO proper serializations
         return jobs_serialized
 
-    def get_job(self, job_id) -> Job:
+    def get(self, job_id) -> Job:
         # TODO validate input
         job = repo.get_job(job_id)
         job_serialized = {
@@ -39,7 +45,7 @@ class JobService:
         } # TODO proper serialization
         return job_serialized
     
-    def add_job(self, script_name, frequency) -> Job:
+    def add(self, script_name, frequency) -> Job:
         # TODO validate inputs
 
         job_serialized = {}
@@ -65,7 +71,7 @@ class JobService:
 
         return job_serialized
     
-    def update_job(self, job_id, new_frequency) -> Job:
+    def update(self, job_id, new_frequency) -> Job:
         # TODO validate inputs
 
         job = repo.get_job(job_id) # TODO error handling
@@ -82,12 +88,31 @@ class JobService:
             "status": "active",
         } # TODO proper serialization
     
-    def cancel_job(self, job_id):
+    def cancel(self, job_id):
         job = repo.get_job(job_id)
         if job:
             # scheduler_service.remove(job) # TODO make this work
             repo.delete_job(job)
 
+            self.unschedule(job)
+
             return jsonify({'message': 'Job deleted successfully'}), 200 # TODO error handling
         else:
-            return
+            return jsonify({'error': 'Job not found'}), 404
+
+    def schedule(self, job: Job): # TODO return something for error handling
+        return scheduler.add_job(
+            func=execute_script,
+            trigger=CronTrigger.from_crontab(job.frequency),
+            args=[job.script_name],
+            id=job.script_name,
+        )
+
+    def reschedule(self, job: Job): # TODO return something for error handling
+        return scheduler.reschedule_job(
+            job_id=job.script_name,
+            trigger=CronTrigger.from_crontab(job.frequency)
+        )
+    
+    def unschedule(self, job: Job): # TODO return something for error handling
+        scheduler.remove_job(job_id=job.script_name)
