+++ app/models/job.py
@@ -1,5 +1,5 @@
 from enum import Enum
-from sqlalchemy import String, Integer
+from sqlalchemy import String
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.extensions import db
